+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All Notable changes to `Slytherin` will be documented in this file
 
 ### Added
 - `Middleware` component
+- `Application::handle` and `Application::toResponse` methods
 
 ## [0.4.3](https://github.com/rougin/slytherin/compare/v0.4.2...v0.4.3) 2016-02-19